+++ engine/Components/print_details.py
@@ -1,5 +1,4 @@
-"""
-Insert
+"""_summary_
 """
 from engine_logging.logging_config import logger
