--- engine/Logging/__init__.py
@@ -1,1 +1,0 @@
-# EMPTY FILE - __init__.py
