+++ engine/Parser/Checks/CheckForIllegalCharacters.py
@@ -1,6 +1,6 @@
-from Components.RegularExpression import validate_input
-from Components.Constants import *
-from Logging.logging_config import logger
+from components.regular_expression import validate_input
+from components.constants import *
+from engine_logging.logging_config import logger
 
 
 class IllegalCharactersException(Exception):
