+++ engine/engine_logging/logging_config.py
@@ -1,3 +1,5 @@
+""" Insert 
+"""
 import logging
 import os
 
@@ -12,15 +14,15 @@ if not os.path.exists(log_folder):
 	os.makedirs(log_folder)
 
 if os.path.exists(log_file):
-	for i in range(MAX_LOG_FILES - 1, 0, -1):  # starts from 9 moving down to 0
-		current_file = f'{log_file}.{i}'
-		next_file = f'{log_file}.{i + 1}'
+    for i in range(MAX_LOG_FILES - 1, 0, -1):  # starts from 9 moving down to 0
+        current_file = f'{log_file}.{i}'
+        next_file = f'{log_file}.{i + 1}'
 
-		if os.path.exists(current_file):
-			if os.path.exists(log_file + '.' + str(MAX_LOG_FILES)):
-				os.remove(log_file + '.' + str(MAX_LOG_FILES))  # this will be next_file
-			os.rename(current_file, next_file)
-	os.rename(log_file, f'{log_file}.1')
+        if os.path.exists(current_file):
+            if os.path.exists(log_file + '.' + str(MAX_LOG_FILES)):
+                os.remove(log_file + '.' + str(MAX_LOG_FILES))  # this will be next_file
+            os.rename(current_file, next_file)
+    os.rename(log_file, f'{log_file}.1')
 
 
 # Configure the logger
